+++ .idea/workspace.xml
@@ -1,0 +1,133 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="AutoImportSettings">
+    <option name="autoReloadType" value="SELECTIVE" />
+  </component>
+  <component name="ChangeListManager">
+    <list default="true" id="1e72d64a-81c4-4828-a95a-1fe0f0380a19" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/src/test/test_diagram.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/test_diagram.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/test_entity.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/test_entity.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/test_imports.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/test_imports.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/test_relation.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/test_relation.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/test_uml_method.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/test_uml_method.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/umleditor/mvc_controller/serializer.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/umleditor/mvc_controller/serializer.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/umleditor/mvc_model/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/umleditor/mvc_model/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/umleditor/mvc_model/custom_exceptions.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/umleditor/custom_exceptions.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/umleditor/mvc_model/diagram.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/umleditor/mvc_model/diagram.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/umleditor/mvc_model/entity.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/umleditor/mvc_model/entity.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/umleditor/mvc_model/relation.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/umleditor/mvc_model/relation.py" afterDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="MarkdownSettingsMigration">
+    <option name="stateVersion" value="1" />
+  </component>
+  <component name="ProjectColorInfo">{
+  &quot;associatedIndex&quot;: 5
+}</component>
+  <component name="ProjectId" id="2e1YiKOl2ntYQ7ogW4DrrkgwBVo" />
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent">{
+  &quot;keyToString&quot;: {
+    &quot;RunOnceActivity.OpenProjectViewOnStart&quot;: &quot;true&quot;,
+    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
+    &quot;git-widget-placeholder&quot;: &quot;coverageFixes&quot;,
+    &quot;node.js.detected.package.eslint&quot;: &quot;true&quot;,
+    &quot;node.js.detected.package.tslint&quot;: &quot;true&quot;,
+    &quot;node.js.selected.package.eslint&quot;: &quot;(autodetect)&quot;,
+    &quot;node.js.selected.package.tslint&quot;: &quot;(autodetect)&quot;,
+    &quot;nodejs_package_manager_path&quot;: &quot;npm&quot;,
+    &quot;settings.editor.selected.configurable&quot;: &quot;org.jetbrains.plugins.github.ui.GithubSettingsConfigurable&quot;,
+    &quot;vue.rearranger.settings.migration&quot;: &quot;true&quot;
+  }
+}</component>
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\Users\kaitl\PycharmProjects\2024sp-420-PythonA\src\umleditor" />
+      <recent name="C:\Users\kaitl\PycharmProjects\2024sp-420-PythonA\src" />
+    </key>
+  </component>
+  <component name="SharedIndexes">
+    <attachedChunks>
+      <set>
+        <option value="bundled-python-sdk-5a2391486177-2887949eec09-com.jetbrains.pycharm.pro.sharedIndexes.bundled-PY-233.13763.11" />
+      </set>
+    </attachedChunks>
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
+  <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="1e72d64a-81c4-4828-a95a-1fe0f0380a19" name="Changes" comment="" />
+      <created>1711073677309</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1711073677309</updated>
+      <workItem from="1711073680164" duration="2170000" />
+      <workItem from="1711214320527" duration="848000" />
+      <workItem from="1711464876084" duration="1316000" />
+      <workItem from="1711466251009" duration="1785000" />
+      <workItem from="1711768620666" duration="149000" />
+      <workItem from="1711808703875" duration="3889000" />
+      <workItem from="1711849532212" duration="5720000" />
+      <workItem from="1711900906631" duration="1819000" />
+      <workItem from="1711940503308" duration="2866000" />
+      <workItem from="1712075391446" duration="10696000" />
+      <workItem from="1712624332409" duration="313000" />
+      <workItem from="1712708147662" duration="781000" />
+      <workItem from="1712758149664" duration="1292000" />
+      <workItem from="1712759522269" duration="12168000" />
+      <workItem from="1712798052257" duration="1413000" />
+      <workItem from="1712975032447" duration="3529000" />
+      <workItem from="1713210093369" duration="25189000" />
+      <workItem from="1713293579760" duration="122000" />
+      <workItem from="1713302467404" duration="6035000" />
+      <workItem from="1713372116802" duration="7060000" />
+      <workItem from="1713392564049" duration="13349000" />
+      <workItem from="1713454506234" duration="8475000" />
+      <workItem from="1713484680101" duration="17702000" />
+      <workItem from="1713535576766" duration="5902000" />
+      <workItem from="1713573876082" duration="5656000" />
+      <workItem from="1713705166677" duration="6449000" />
+      <workItem from="1713742489819" duration="7739000" />
+      <workItem from="1713754815078" duration="3283000" />
+      <workItem from="1713985392538" duration="1647000" />
+      <workItem from="1714075201122" duration="8831000" />
+      <workItem from="1714155430014" duration="414000" />
+    </task>
+    <servers />
+  </component>
+  <component name="TypeScriptGeneratedFilesManager">
+    <option name="version" value="3" />
+  </component>
+  <component name="Vcs.Log.Tabs.Properties">
+    <option name="TAB_STATES">
+      <map>
+        <entry key="MAIN">
+          <value>
+            <State />
+          </value>
+        </entry>
+      </map>
+    </option>
+  </component>
+  <component name="XDebuggerManager">
+    <breakpoint-manager>
+      <breakpoints>
+        <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
+          <url>file://$PROJECT_DIR$/src/test/test_relation.py</url>
+          <line>5</line>
+          <option name="timeStamp" value="2" />
+        </line-breakpoint>
+      </breakpoints>
+    </breakpoint-manager>
+  </component>
+</project>
