+++ src/test/test_uml_method.py
@@ -1,4 +1,5 @@
-from umleditor.mvc_model import UML_Method, CustomExceptions
+from umleditor.mvc_model.entity import UML_Method
+from umleditor.custom_exceptions import CustomExceptions
 import pytest
 
 def test_create_method( ):
