+++ src/umleditor/mvc_model/__init__.py
@@ -1,7 +1,6 @@
 import sys
 sys.path.append('src/umleditor/mvc_model')
 
-from .custom_exceptions import *
 from .diagram import *
 from .entity import *
 from .relation import *
