+++ src/umleditor/mvc_model/diagram.py
@@ -1,6 +1,6 @@
-from .entity import Entity, UML_Method
+from .entity import Entity
 from .relation import Relation
-from .custom_exceptions import CustomExceptions
+from umleditor.custom_exceptions import CustomExceptions
 
 class Diagram:
     #Singleton Design Pattern
